synchronize cards with tv-series api
