+++ src/cardStyles.css
@@ -1,3 +1,10 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,400;0,500;1,200&family=Roboto:ital,wght@0,100;0,300;1,300;1,700&display=swap');
+
+:root {
+  --font-poppins: 'Poppins', sans-serif;
+  --font-roboto: 'Roboto', sans-serif;
+}
+
 .main-sec {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
@@ -9,11 +16,30 @@
 .card {
   background-color: #fff;
   border: 1px solid #ccc;
-  border-radius: 5px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  padding: 20px;
+  border-radius: 6px;
+  box-shadow: 0 0 12px rgba(0, 0, 0, 0.1);
+  padding: 6px;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+}
+
+.card-subtitle {
+  font-family: var(--font-poppins);
+  font-size: 0.9rem;
+  font-weight: bold;
+  color: #000;
+  display: flex;
+  align-items: center;
+  height: 2em;
+  width: 200px;
+  justify-content: space-between;
+}
+
+.likes {
+  display: flex;
+  align-items: center;
+  justify-self: flex-end;
+  height: 42px;
 }
